+++ src/hooks/earn/useEarnRelatedMarkets.ts
@@ -1,14 +1,10 @@
-import { getOpportunityRelatedMarket } from '@/app/lib/getOpportunityRelatedMarket';
 import { useQuery } from '@tanstack/react-query';
+
+import {
+  earnRelatedMarketsQueryKey,
+  fetchEarnRelatedMarkets,
+} from '@/app/lib/earn/earnQueries';
 import { FIVE_MINUTES_MS } from '@/const/time';
-
-export const earnRelatedMarketsQueryKey = (slug: string) =>
-  ['earn-related-markets', slug] as const;
-
-export const fetchEarnRelatedMarkets = async (slug: string) => {
-  const result = await getOpportunityRelatedMarket(slug);
-  return result.data;
-};
 
 export const useEarnRelatedMarkets = (slug: string) => {
   return useQuery({
