+++ src/hooks/earn/useEarnTopOpportunities.ts
@@ -1,6 +1,9 @@
 import type { UseQueryResult } from '@tanstack/react-query';
 import { useQuery } from '@tanstack/react-query';
-import { getOpportunitiesTop } from 'src/app/lib/getOpportunitiesTop';
+import {
+  earnTopOpportunitiesQueryKey,
+  fetchEarnTopOpportunities,
+} from '@/app/lib/earn/earnQueries';
 import { FIVE_MINUTES_MS } from 'src/const/time';
 import type { EarnOpportunityWithLatestAnalytics } from 'src/types/jumper-backend';
 import type { Hex } from 'viem';
@@ -12,17 +15,6 @@ export type Result = UseQueryResult<
   EarnOpportunityWithLatestAnalytics[],
   unknown
 >;
-
-export const earnTopOpportunitiesQueryKey = (address?: Hex) =>
-  ['earn-top-opportunities', address] as const;
-
-export const fetchEarnTopOpportunities = async (address?: Hex) => {
-  const result = await getOpportunitiesTop(address);
-  if (!result.ok) {
-    throw result.error;
-  }
-  return result.data.data;
-};
 
 export const useEarnTopOpportunities = ({}: Props): Result => {
   const address: Hex | undefined = useAccountAddress();
