+++ src/providers/ReactQueryProvider.tsx
@@ -1,29 +1,29 @@
 'use client';
 
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-// import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import React from 'react';
+import type { QueryClient } from '@tanstack/react-query';
+import { QueryClientProvider } from '@tanstack/react-query';
+
+import { makeQueryClient } from '@/app/lib/makeQueryClient';
+
+let browserQueryClient: QueryClient | undefined;
+
+const getQueryClient = () => {
+  if (typeof window === 'undefined') {
+    return makeQueryClient();
+  }
+
+  if (!browserQueryClient) {
+    browserQueryClient = makeQueryClient();
+  }
+
+  return browserQueryClient;
+};
 
 export function ReactQueryProvider({ children }: React.PropsWithChildren) {
-  const [client] = React.useState(
-    new QueryClient({
-      defaultOptions: {
-        queries: {
-          enabled: true,
-          staleTime: 3_600_000,
-          refetchInterval: false,
-          refetchIntervalInBackground: false,
-          refetchOnWindowFocus: true,
-          refetchOnReconnect: true,
-          refetchOnMount: true,
-          retryOnMount: true,
-        },
-      },
-    }),
-  );
+  const queryClient = getQueryClient();
 
   return (
-    <QueryClientProvider client={client}>
+    <QueryClientProvider client={queryClient}>
       {children}
       {/* <ReactQueryDevtools initialIsOpen={false} /> */}
     </QueryClientProvider>
